Move code over into `assessment.py`.
Introduces the `DLCalculationAssessment` class. This class can
co-exist with the current implementation of `Assessment`, the methods
of which are brand new and a bit "cleaner", but can't be directly used
to substitute what existed in `DL_calculation.py`. The
`DLCalculationAssessment` contains surgically extracted code from
`DL_calculation.py` that was slowly and carefully refactored with the
intention of keeping the input/output behavior unchanged. Future
changes could include introducing sub-classes to separate cases for
water network and other kinds of assessments (to remove the current
`is_water` if/else logic).
