+++ .pylintrc
@@ -102,6 +102,9 @@ disable=
     format,
     # We anticipate #3512 where it will become optional
     fixme,
+    arguments-differ,
+    arguments-renamed,
+    else-if-used,
 
 [REPORTS]
 
@@ -422,7 +425,7 @@ max-spelling-suggestions=2
 [DESIGN]
 
 # Maximum number of arguments for function / method
-max-args=12
+max-args=15
 
 # Maximum number of locals for function / method body
 max-locals=50
@@ -543,7 +546,7 @@ bad-functions=map,input
 [REFACTORING]
 
 # Maximum number of nested blocks for function / method body
-max-nested-blocks=8
+max-nested-blocks=9
 
 # Complete name of functions that never returns. When checking for
 # inconsistent-return-statements if a never returning function is called then
